requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
google-genai==1.59.0
gunicorn==21.2.0; sys_platform != 'win32'